% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diag_codes.R
\name{get_diag_codes}
\alias{get_diag_codes}
\title{retrieves diagnosis (dx) codes matching a wildchar}
\usage{
get_diag_codes(diag_codes = "__NULL__", api_function = "get_diag_codes",
  ...)
}
\arguments{
\item{diag_codes}{a string with the requested pattern for diagnosis code}

\item{api_function}{a variable indicating the name of the function. It is unlikely that you'll ever
need to change the value of this parameter and it is best left as-is}
}
\value{
a vector with the diagnosis codes
}
\description{
The \code{get_diag_codes} function processes extracts diagnosis codes from
the dim_dx table matching the pattern passed in the function
}
\examples{

# extract diagnosis codes starting with E11
get_diag_codes("E11*")
}
