% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_query.R
\name{process_query}
\alias{process_query}
\title{a wrapper (function) to execute an API function}
\usage{
process_query(api_function, get_stats = "FALSE", ...)
}
\arguments{
\item{api_function}{a character value with the name of the api function}

\item{get_stats}{a boolean variable indicating whether query statistics should be printed}
}
\value{
a data table with the results of the query
}
\description{
The \code{process_query} function processes parameters passed by any of the API function
by converting them first to JSON format and then normalising the string with an URLencode command.
It is not intended to be run directly by the end user and is instead used by the package to wrap
API calls.
}
