% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities}
\alias{utilities}
\alias{test_connection}
\alias{reset_connection}
\alias{get_table_stats}
\title{helper utilities for the RXDS API for Truven Health Research Database}
\usage{
test_connection()

reset_connection()

get_table_stats(RStats, kdbStats, startTime, endTime)
}
\arguments{
\item{RStats}{a data table with R connection statistics}

\item{kdbStats}{a data table with KDB+ connection statistics}

\item{startTime}{a character value with start time}

\item{endTime}{a character value with end time}
}
\value{
A message confirming the status of the connection attempt

None

a data table with R Stats and KDB+ Stats
}
\description{
The \code{test_connection} function initiates a connection to a target R Server
using the servername and port number. The user simply needs to execute the function
\code{test_connection()} which will prompt for the server and port information. The function
then attempts to establish a connection. The user will get a message confirming whether the
connection was successful or otherwise.

Important: The user should never have to run this function except for debugging purposes. Each
function in the API automatically executes this function to test for connectivity. It is intended
to merely indicate the existence of a valid server:port connection.

When this function is executed successfully, it sets 2 global variables:
rxds_api_r_kdb_server: The server name; and
rxds_api_r_kdb_port: The port number of the server

In the event that the connection is not successful, the user may re-execute the function or
contact support for further information.

This function does not accept any parameter.

The \code{reset_connection} function resets connection details for a target R server.
When a connection to a target server is established, the system sets 2 global variables (rxds_api_r_kdb_server
and rxds_api_r_kdb_port). This function resets the value of the variables and when the user
attempts to run a new function, it prompts the user for the relevant server/port information

In the event that the connection is not successful, the user may re-execute the function or
contact support for further information.

This function does not accept any parameter.

The \code{get_table_stats} combines data from R and KDB+ execution statistics
This is a 'helper' function for other functions in the package and is not intended to be execute
directly by the end user
}
\examples{

}
